+++ kelk/env/src/export.rs
@@ -27,10 +27,11 @@ extern "C" fn deallocate(ptr_u64: u64) {
 
 /// do_instantiate should be wrapped in an external "C" export,
 /// containing a contract-specific function as arg.
-pub fn do_instantiate<E: Encode>(_instantiate_fn: &dyn Fn(Context) -> Result<(), E>) -> u32 {
-    //let ctx = make_context();
-    //instantiate_fn(ctx.as_ref());
-    0
+pub fn do_instantiate<'a, D: Decode<'a>, E: Encode>(
+    instantiate_fn: &dyn Fn(Context, D) -> Result<(), E>,
+    msg_ptr: u64,
+) -> u64 {
+    do_execute(instantiate_fn, msg_ptr)
 }
 
 /// do_process_msg should be wrapped in an external "C" export,
@@ -39,13 +40,7 @@ pub fn do_process_msg<'a, D: Decode<'a>, E: Encode>(
     process_msg_fn: &dyn Fn(Context, D) -> Result<(), E>,
     msg_ptr: u64,
 ) -> u64 {
-    let ptr = Pointer::from_u64(msg_ptr);
-    let buf = unsafe { ptr.to_slice() };
-    let msg = minicbor::decode(buf).expect("Decoding failed");
-    let ctx = make_context();
-    let res = process_msg_fn(ctx.as_ref(), msg);
-
-    result_to_ptr(res)
+    do_execute(process_msg_fn, msg_ptr)
 }
 
 /// do_query should be wrapped in an external "C" export,
@@ -54,16 +49,18 @@ pub fn do_query<'a, D: Decode<'a>, R: Encode, E: Encode>(
     query_fn: &dyn Fn(Context, D) -> Result<R, E>,
     msg_ptr: u64,
 ) -> u64 {
+    do_execute(query_fn, msg_ptr)
+}
+
+fn do_execute<'a, D: Decode<'a>, R: Encode, E: Encode>(
+    func: &dyn Fn(Context, D) -> Result<R, E>,
+    msg_ptr: u64,
+) -> u64 {
     let ptr = Pointer::from_u64(msg_ptr);
     let buf = unsafe { ptr.to_slice() };
     let msg = minicbor::decode(buf).expect("Decoding failed");
     let ctx = make_context();
-    let res = query_fn(ctx.as_ref(), msg);
-
-    result_to_ptr(res)
-}
-
-fn result_to_ptr<E: Encode>(res: E) -> u64 {
+    let res = func(ctx.as_ref(), msg);
     let mut vec = kelk_lib::alloc::vec::Vec::new();
     minicbor::encode(res, &mut vec).expect("Encoding failed");
 
@@ -95,6 +92,34 @@ mod tests {
     //     // Should panic here, because the pointer is freed before
     //     deallocate(ptr);
     // }
+
+    #[wasm_bindgen_test]
+    fn test_instantiate() {
+        let msg_data = vec![0x00]; // http://cbor.me/?bytes=8100
+        let msg_ptr = Pointer::release_buffer(msg_data);
+
+        let res_ptr = do_instantiate(
+            &|_: Context, _: i32| -> Result<(), i32> { Ok(()) },
+            msg_ptr.as_u64(),
+        );
+
+        let res_data = unsafe { Pointer::from_u64(res_ptr).to_slice() };
+        assert_eq!(res_data, vec![0x82, 0x00, 0x80]); // Result::Ok(()) -> http://cbor.me/?bytes=820080
+    }
+
+    #[wasm_bindgen_test]
+    fn test_instantiate_error() {
+        let msg_data = vec![0x00]; // http://cbor.me/?bytes=8100
+        let msg_ptr = Pointer::release_buffer(msg_data);
+
+        let res_ptr = do_instantiate(
+            &|_: Context, _: i32| -> Result<(), i32> { Err(0x0e) },
+            msg_ptr.as_u64(),
+        );
+
+        let res_data = unsafe { Pointer::from_u64(res_ptr).to_slice() };
+        assert_eq!(res_data, vec![0x82, 0x01, 0x0e]); // Result::Err(0x0e) -> http://cbor.me/?bytes=82010e
+    }
 
     #[wasm_bindgen_test]
     fn test_process_msg() {