+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Defining Memory Pointer for managing memory allocation in WASM's linear memory.
 - Implementing `deallocate` API to free the allocated memory
 - Implementing Query API
+- Implementing Instantiate API
 - Updating Storage API for primitive types
 - Storage BST (Binary Search Tree) for lib crate